pyproject.toml
[tool.poetry]
name = "slider"
version = "0.1.0"
description = ""
authors = ["copyrite"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
steamleaderboards = "^0.0.3"
python-valve = "^0.2.1"
more-itertools = "^9.0.0"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
